src/main/java/com/springsecurity/rbac/springsecurityrbac/service/UserService.java
package com.springsecurity.rbac.springsecurityrbac.service;

import com.springsecurity.rbac.springsecurityrbac.dto.UserDto;
import com.springsecurity.rbac.springsecurityrbac.entity.User;
import com.springsecurity.rbac.springsecurityrbac.exception.UserAlreadyExistException;
import com.springsecurity.rbac.springsecurityrbac.mapper.UserMapper;
import com.springsecurity.rbac.springsecurityrbac.repository.UserRepository;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.time.LocalDateTime;
import java.util.Collection;
import java.util.Optional;

@Service
@Transactional
public class UserService {
    private UserRepository userRepository;
    private PasswordEncoder passwordEncoder;


    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    public UserDto createUser(UserDto userDto) throws UserAlreadyExistException {
        if (userRepository.existsByEmail(userDto.getEmail())) {
            throw new UserAlreadyExistException("User with " + userDto.getEmail() + " already exist!");
        }
        User user = UserMapper.toUser(userDto);
        user.setSpecialPrivileges(false);
        user.setCreatedAt(LocalDateTime.now());
        user.setPassword(passwordEncoder.encode(userDto.getPassword()));
        return UserMapper.toUserDto(userRepository.save(user));
    }

    public Collection<UserDto> findAll() {
        return UserMapper.toUserDtos(userRepository.findAll());
    }

    public UserDto findByEmail(String username) throws UsernameNotFoundException {
        Optional<User> optionalUser = userRepository.findByEmail(username);
        return UserMapper.toUserDto(optionalUser.orElseThrow(
                () -> new UsernameNotFoundException("User with email " + username + " not found")
        ));
    }

    public UserDto deleteByEmail(String email) throws UsernameNotFoundException {
        Optional<User> optionalUser = userRepository.findByEmail(email);
        User user = optionalUser.orElseThrow(
                () -> new UsernameNotFoundException("User with email " + email + " not found")
        );
        userRepository.delete(user);
        return UserMapper.toUserDto(user);
    }
}
